swiftrepl: add a -o (once) option
Add a -o (once) option that doesn't loop over containers forever but
finishes up and moves to the next container when it's done. Especially
helpful with sync_deletes().

Also, join all threads in the main thread instead of while True:
sleep(10). This in combination with the above means that swiftrepl
actually exits when done with all containers.

Change-Id: I265e81c341568373d952e69804cb7611928bd6d0