swiftrepl: add an error message for IncompleteSend
If an object on the source had a mismatching size between the container
listing and the actual file (which, sadly, happens in Swift), the code
would raise an IncompleteSend() and stop processing that container.

Ideally, this should be handled gracefully and continue processing the
container. Until then though, add an error message which informs us of
the container and the object so that we can fix it up on the source.

Change-Id: I86992a3b5e33b1c05db17e8276a80809ff8a8540