Tools to compare compiled puppet catalogs.
This set of tools (a base bash/ruby catalog compiler that supports
different puppet versions, and a python controller of the whole process)
allows to compile our whole puppet catalog or a subset of it, once the
yaml facts files are uploaded in its external/var directory.

It generates compilation logs, and diff files as well as an html
representation of the result, for easy consulting.

Also, a base vagrant configuration with provisioning of a working box
for local compilation and evaluation has been added so that you can
easily check the compilation status.

Change-Id: Id86239e0d5e430b23bd4b670dec383bc57877041
Signed-off-by: Giuseppe Lavagetto <glavagetto@wikimedia.org>