compare-puppet-catalogs: new puppet catalog diff
As puppet catalog diff tool has been updated and completely changed its
output format, a corresponding change in our catalog comparator was
needed as well.

Now we get a (cleaner) json file as output and use it to compute and
build differences on the go; this makes also easier to give a somewhat
terser html output (dividing diff by resource).

Change-Id: Ia12d8fccb9054c095919dcd36b411855962181d3
Signed-off-by: Giuseppe Lavagetto <glavagetto@wikimedia.org>