Use MariaDB event scheduler for labsdb.
Start by checking for and killing threads that:

- Are normal labs users /^[usp][0-9]+/
- Have been in Sleep state for over 300s
- Have an open, uncommitted transaction in InnoDB

This sort of thing happens when people use the RDBMS as locking
mechanism for something external, or don't issue a COMMIT with a
connection pool or some sort, or client connections fail in various
obscure ways that don't timeout the TCP socket. The other safety
net is wait_timeout=28800, which is, necessarily, too far off to
help with uncommitted transactions.

Change-Id: Id792f752339d46cdc8124dc471b0d8d0897fad5a