WIP: Cleanup the Sanitarium
We already set triggers on tables for the nulling out fields
replicated to labs; extend that to a cleaner whitelist approach
that does not care when new tables and wikis are created.

Make it easier to resync the labs slaves by using activity
logging along side normal replication, designed along the lines
of Tungsten Replicator (without the java bit).

Probably also relevant to the impending replication of some Event
Logging tables to labs, pending discussion.

Change-Id: I4d060cce995380709caf5facada15c57bc3fc5a1