Add SQL for the creation of heartbeat tables
This sql files creates the database heartbeat_p
and the view heartbeat that allows monitoring the lag
on labsdb hosts, as compared with its masters. This is a long-
time requested feature.

This does not setup pt-heartbeat, and it just assumes it is
there, as it is not filtered but setup in production.

This is not integrated with maint-replicas.pl and I am
a horrible human person for that. But at least it is on a repo.

In order to recreate, run:
$ mysql -h [labs-host] < heartbeat-views.sql

The usage is documented on:
https://wikitech.wikimedia.org/wiki/Help:Tool_Labs/Database#Identifying_lag

MariaDB currently has a bug by which it doesn't allow MICROSECOND
resolution. This will be fixed when MariaDB fixes it:
https://mariadb.atlassian.net/browse/MDEV-9175

Bug: T71463
Change-Id: Ifaad9e187704943d71c98d6e835a8d33da8280a0