swiftrepl: fix destination container listing limit
while testing the latest version of swiftrepl I came across an http error from
swift "412 Precondition failed". As it turns out swiftrepl was asking container
listings with a limit of 20000 but default maximum is 10000.

My understanding of what happened is this: 'limit' variable assignment was
removed in I148b0c209 from the function, however a global 'limit' variable was
introduced in I464b6a68. In I66a04cef1 there's been a cleanup of variables but
'limit' was re-introduced outside the 'while True' loop, not inside.

Bug: T125791
Change-Id: I17614830f79e7806b1d253df8209473e89f9fb41