DBtools: add script to check external storage
Usage: ${0} -s START_TIMESTAMP -e END_TIMESTAMP -r REVISION_DB
       -o ORIGIN_DB -t TARGET_DB -c CLUSTER -d DATABASES_FILE

Example: ${0} -s 20160131235959 -e 20160201020000
         -r dbstore1002.eqiad.wmnet -o es1019.eqiad.wmnet
         -t es2019.codfw.wmnet -c 25 -d databases.txt

External Storage check script to ensure that the content of a group of
blobs id and text content is the same between the ORIGIN_DB server and
the TARGET_DB one for all revisions created between START_TIMESTAMP
and END_TIMESTAMP for the given CLUSTER.
The revision table to get the timestamps is checked on the REVISION_DB
server.
DATABASES_FILE is expected to be a file with all the databases to be
checked, one per line.
Timestamps must be in the format used by MediaWiki: YYYYmmddHHMMSS
All parameters are mandatory.

Bug: T130702
Change-Id: I8df901bbed858117393b182920bae52e07550ca2