Remove /.pep8 file: no more needed
All the configuration has been moved to tox.ini [flake8] the only bit of
interest is "filename" which was set to get flake8 to pass on files that
do not ends with '.py'.

From them:

 geturls: got renamed to geturls.py
 swiftcleaner: is legacy, unlikely to be fixed
 profiler-to-carbon: udpprofile is no more

The rest of the ignore are not needed, and we have a list in tox.ini
already.

Change-Id: Iddc2fc9af402a6dacf54381f27bda5ae212ee706