Add script to generate mysql client-only .deb package
This package is needed because regular client library
is not compatible with proper TLS support, failing to connect
to production servers with an 'unknown SSL error'.

This should only be installed on client-only hosts, as the
server package contains the client too. Nothing wrong will happen
if they both are installed, except a waste of space (and using
the wrong mysql client version).

Bug: T157702

Change-Id: Ie2bf362774e4dbad168f89ce21675f84f0df04d6