generate_dsns_table.sh: Add new script for dsns.
The dsns tables are used by pt-table-checksum as another method to
discover slaves instead of "show processlist".

We are using this method in order to tell pt-table-checksum which slaves
it should look for, so we can discard dbstores and sanitarium hosts
which might be delayed and then would stop the check forever.

This quick and small shell script just parses the output from *.hosts
files to identify the slaves, remove the dbstore or sanitarium ones, as
well as the primary master and builds the syntax to run on mysql to
populate the dsns tables.

The dsns tables live under the dsns database on Tendril hosts.
There is one table per shard at the moment.

Bug: T154485
Change-Id: Iffb53744ca71c6a3f6c743f22e7ef80b121223a7