mariadb.service: Set start/stop timeout to infinity
It currently timeouts with the default value, and says the unit
failed even if it stopped cleanly, or in certain cases, kill
the server and corrupt the data.

Even 30 minutes may not be enough for a server to legitimate stop
or start, if it has to recover after crashing- set it to infinity.

Also build jessie 10.1 packages with systemd support, too.

Change-Id: Ie62f522aeb0eee0fdfdf58788476defa3e0e2b75