s1.hosts: Add db2092 to s1
db2092 is now a spare. Move it to s1.
In order not to have db2033 as a candidate master
and as unique slave for codfw x1 as it has a faulty BBU T184888
 I would like to do the following:

Move db2092 (spare 512G host - T170662) to s1 as API
Move db2069 (160G host) from s1 API to x1 slave+candidate master

Bug: T191275
Bug: T170662
Change-Id: Id7a76d06c99838d3c7211b82700bfb5c181d9353