mariadb: Add main_tables.txt, collection of main mediawiki tables
This files contrains a representative set of tables from a
mediawiki installation, and its primary keys. This allows to do
a quick consistency check without actually checking every single
table. In the future this list may be used for an automated run
of compare.py to detect quickly consistency issues on servers.

Bug: T104459
Change-Id: I4d802235ab624a195c196ee6537f29e608ab74cf