dbtools: check_tables.sh
This is kind of a oneliner I have been using recently to check out for
change_table drifts and can be used with any other table.
It now uses "section" and compare.py.

Again, this is just a oneliner to be run from neodymium, it doesn't
pretend to be fancy or anything. It was put together to make checks
faster.

However it can be cronned if needed to start checking some tables
periodically, tables like change_tag or tag_summary which have had very
big issues in the past as they didn't have a PK

Bug: T104459
Change-Id: I9682fd22c6604c113a99d79b1bca6fe06a19f24c