Initial commit of quotereviewer
Quotereviewer is a tool to automatically parse Dell quotes, and match
their contents against a manually-maintained site policy, shipped in a
YAML file.

The policy currently can contain three different lists of SKUs:
• "required": SKUs that should always be present. An example would be
  "No Operating System" or "On-Site Installation Declined".
• "blacklist": SKUs that should absolutely not be present. For example,
  "No Rack Rails or Cable Management Arm".
• "optional": SKUs that can be present, but are not required to be.

Anything that is not in either of these three lists is considered
unknown and is emitted as a warning, with the expectation that they
should immediately be added to one of those three lists. Only quotes
with all of their SKUs known are accepted.

An example output would be:
  The following blacklisted SKUs are present:
    • 350-BBKQ Quick Sync 2 (At-the-box mgmt)
    • 770-BBHJ No Rack Rails or Cable Management Arm

Change-Id: If1701905c1fe06118e0419722c2ac142e5fa5e82