quotereviewer: move Declined Remote Consulting Service back to optional
I had previously added 973-2426 'Declined Remote Consulting Service'
to the required section, however that is an option that now only shows
up when I use the Premier portal to generate quotes. It no longer
(according to our team) shows up when Dell generates an R440 quote going
forward. To quote the email from Dell: "You are correct on the quote
there is no remote consulting. Our tool differs from the online tool on
certain features, in this instance we don’t even have the option to
select remote consulting, we would have to add the SKU to the quote
manually on our side so our services team could be flagged correctly and
engage the customer for services."

Since it seems this was included on previous quotes, I've moved it to
the optional section, undoing my addition of it to the required section.

Change-Id: I8abab7144349d89651ea0039bc3fd1d85991058a