mariadb: Add mbstream to path
https://gerrit.wikimedia.org/r/c/operations/puppet/+/599596 changed
transfer.py to use $PATH for finding the mariadb tools. The current
packages do not make mbstream available in $PATH, so the puppet CR
temporarily added a symlink.

This change ensures that mbstream will be on $PATH, so the workaround
can be removed once new packages are built and pushed.

Change-Id: Ibf86e589c458bcba424675d6d6bc8fed553364d4