updating sku listing
i had to move some 'required' skus down to the optional area, as there
are differing skus for the R740xd2.  example: basic is moved odwn for
the two different basic support skus.  the 'no support' is still
blacklisted, and any basic sku that shows up (that isnt 36 months) will
show up as unknown sku so we know to investigate.

Change-Id: I752c92a4531e78ee0e2c67c620e8f4b7c4907df8