dbtools: make mariadb service Wants prometheus-mysqld-exporter
If prometheus-mysqld-exporter is available, we want to start it up
after mysql/mariadb is started. This change to the service file
achieves that.

This sort of relationship is relatively weak - if the PME service is
absent or fails, mariadb will still be started OK.

The same change is made to both the single and multi-instance service
files.

A change made elsewhere in puppet will ensure, in addition, that the
prometheus-mysqld-exporter service is stopped/restarted when mariadb
is.

Bug: T289488
Change-Id: I6bb04835af41401c8939568c4ae9a4625d23e23b