clinic-duty: add coverage for work.gcalendarLink()
* Fix TODO about global state and inject into Message class
  accordingly.

* Move DOM use to caller, reducing gcalendarLink() to just
  the href return.

* Re-enable tests, and use URL API to parse the URL, and assert
  the 'dates' param.

To use the URL API without warnings from ESLint, update to the
latest eslint-config-wikimedia version, and add the missing 'engines'
field to package.json as otherwise it will warn for this constructor
not being available in (now EOL) Node.js 8.

The latest eslint-config-wikimedia version also warns against use
of non-strict assertions in tests, so fix those while at it.

Change-Id: I77780e6dec31b4b43621328d04f9da0928279565