ops-maint-gcal.js: truncate message details
Google will return a 400 response if the total length of the URL path +
query params exceeds 16k.

The "details" param (i.e., message link + text) is now truncated so we
stay within those limits, and a warning is emitted to the console.

Since the length constraint applies post-encoding, and the message may
contain multi-byte code points, the "simplest" way to find the longest
truncation that works is just to bisect.

Bug: T381680
Change-Id: I2b454d092df1a7b42d702b8d4c94c43780608481