reboot_es.sh: Reboot standalone external store
I wrote this sometime ago as at the time we didn't have an upgrade host
cookbook. The current one still doesn't work as these hosts are
standalone.

Just pushing it to the repo so it is somewhere else and it is tracked.

There are many flaws and it is not meant to be used as a only source of
automation for these hosts.

Change-Id: I9abb2fd6a69a5011df4e1f7db2dd71c93ef5002d