report_users.sh: Add check for public role
This is a new role in mariadb 10.11 which gets installed when we install
a new database (something we rarely do).
This role cannot be deleted but also doesn't allow everyone to access a
test db (which we also don't do)

https://jira.mariadb.org/browse/MDEV-5215
https://mariadb.com/docs/server/security/user-account-management/roles/system-users-roles-and-privileges

For now, we can just simply alert when this role has privileges and if
it does, we can just remove them with REVOKE.
By default this role is only for test.* databases and has no users in
it.

Change-Id: Ice95b74104292726416a7fc0cc486cc89f14a5e1